@@ -3,6 +3,7 @@ import React from 'react'
 
 import ReactPaginate from '../../common/paginate/ReactPaginate.jsx'
 import IPsSocketioEventsEmitter from '../../redux/ips/IPsSocketioEventsEmitter.js'
+import CredsSocketioEventsEmitter from '../../redux/creds/CredsSocketioEventsEmitter.js'
 import IPEntryLine from '../presentational/scope/IPEntryLine.jsx'
 import Search from '../../common/search/Search.jsx'
 
@@ -23,6 +24,7 @@ class IPTable extends React.Component {
 		}
 
 		this.ipsEmitter = new IPsSocketioEventsEmitter();
+		this.credsEmitter = new CredsSocketioEventsEmitter();
 
 		this.commentSubmitted = this.commentSubmitted.bind(this);
 		this.handlePageClick = this.handlePageClick.bind(this);
@@ -36,25 +38,25 @@ class IPTable extends React.Component {
 		return (!_.isEqual(nextProps, this.props) || !_.isEqual(this.state, nextState));
 	}
 
-	componentWillReceiveProps(nextProps) {
-		if ((nextProps.ips.page !== this.props.ips.page) || (nextProps.ips.page_size !== this.props.ips.page_size)) {
+	componentDidUpdate(prevProps) {
+		if ((this.props.ips.page !== prevProps.ips.page) || (this.props.ips.page_size !== prevProps.ips.page_size)) {
 			this.props.setLoaded(true);
 		}
 
-		if (nextProps.ips.selected_ips !== this.props.ips.selected_ips) {
+		if (this.props.ips.selected_ips !== prevProps.ips.selected_ips) {
 			this.setState({
-				shownData: nextProps.ips.data,
+				shownData: this.props.ips.data,
 				offsetPage: 0,
-				pageCount: Math.ceil(nextProps.ips.selected_ips / this.props.ips.page_size)
+				pageCount: Math.ceil(this.props.ips.selected_ips / this.props.ips.page_size)
 			});
 		}
 		else {
 			this.setState({
-				shownData: nextProps.ips.data,
-				pageCount: Math.ceil(nextProps.ips.selected_ips / this.props.ips.page_size)
+				shownData: this.props.ips.data,
+				pageCount: Math.ceil(this.props.ips.selected_ips / this.props.ips.page_size)
 			});
 		}
-	}	
+	}
 
 	handlePageClick(page_number) {
 		this.props.setLoaded(false);
