@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types';
 
 import HostPage from '../host_verbose/components/HostVerboseWrapper.jsx'
 
@@ -7,6 +8,7 @@ import HostsSocketioEventsSubsriber from '../redux/hosts/HostsSocketioEventsSubs
 import TasksSocketioEventsSubsriber from '../redux/tasks/TasksSocketioEventsSubsriber'
 import ScansSocketioEventsSubsriber from '../redux/scans/ScansSocketioEventsSubscriber'
 import FilesSocketioEventsSubsriber from '../redux/files/FilesSocketioEventsSubscriber'
+import CredsSocketioEventsSubscriber from '../redux/creds/CredsSocketioEventsSubscriber'
 
 
 class Host extends React.Component {
@@ -24,6 +26,7 @@ class Host extends React.Component {
         this.tasksSubscriber = new TasksSocketioEventsSubsriber(mainStore, project_uuid);
         this.scansSubscriber = new ScansSocketioEventsSubsriber(mainStore, project_uuid);
         this.filesSubscriber = new FilesSocketioEventsSubsriber(mainStore, project_uuid, hostname);        
+        this.credsSubscriber = new CredsSocketioEventsSubscriber(mainStore, project_uuid);
     }
 
     render() {
@@ -37,11 +40,12 @@ class Host extends React.Component {
         this.tasksSubscriber.close();
         this.scansSubscriber.close();
         this.filesSubscriber.close();        
+        this.credsSubscriber.close();    
     }       
 }
 
 Host.contextTypes = {
-    store: React.PropTypes.object
+    store: PropTypes.object
 }
 
 export default Host;
