@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types';
 
 import IPPage from '../ip_verbose/components/MainAccumulatorWrapper.jsx'
 
@@ -7,6 +8,7 @@ import IPsSocketioEventsSubsriber from '../redux/ips/IPsSocketioEventsSubscriber
 import TasksSocketioEventsSubsriber from '../redux/tasks/TasksSocketioEventsSubsriber'
 import ScansSocketioEventsSubsriber from '../redux/scans/ScansSocketioEventsSubscriber'
 import FilesSocketioEventsSubsriber from '../redux/files/FilesSocketioEventsSubscriber'
+import CredsSocketioEventsSubscriber from '../redux/creds/CredsSocketioEventsSubscriber'
 
 
 class IP extends React.Component {
@@ -24,6 +26,7 @@ class IP extends React.Component {
         this.tasksSubscriber = new TasksSocketioEventsSubsriber(mainStore, project_uuid);
         this.scansSubscriber = new ScansSocketioEventsSubsriber(mainStore, project_uuid);
         this.filesSubscriber = new FilesSocketioEventsSubsriber(mainStore, project_uuid, ip_address);        
+        this.credsSubscriber = new CredsSocketioEventsSubscriber(mainStore, project_uuid);
     }
 
     render() {
@@ -37,12 +40,13 @@ class IP extends React.Component {
         this.ipsSubscriber.close();
         this.tasksSubscriber.close();
         this.scansSubscriber.close();
-        this.filesSubscriber.close();        
+        this.filesSubscriber.close();    
+        this.credsSubscriber.close();    
     }       
 }
 
 IP.contextTypes = {
-    store: React.PropTypes.object
+    store: PropTypes.object
 }
 
 export default IP;
