@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types';
 
 import NavigationTabs from './NavigationTabs.jsx'
 
@@ -10,7 +11,7 @@ import ScansSocketioEventsSubsriber from '../redux/scans/ScansSocketioEventsSubs
 import FilesSocketioEventsSubsriber from '../redux/files/FilesSocketioEventsSubscriber'
 import NotificationsSocketioEventsSubscriber from '../redux/notifications/NotificationsSocketioEventsSubscriber'
 import ScopesSocketioEventsSubscriber from '../redux/scopes/ScopesSocketioEventsSubscriber'
-
+import CredsSocketioEventsSubscriber from '../redux/creds/CredsSocketioEventsSubscriber'
 
 class NavigationTabsWrapper extends React.Component {
     constructor(props) {
@@ -29,6 +30,7 @@ class NavigationTabsWrapper extends React.Component {
         this.filesSubscriber = new FilesSocketioEventsSubsriber(mainStore, project_uuid);   
         this.notificationsSubscriber = new NotificationsSocketioEventsSubscriber(mainStore, project_uuid);
         this.scopesSubscriber = new ScopesSocketioEventsSubscriber(mainStore, project_uuid);
+        this.credsSubscriber = new CredsSocketioEventsSubscriber(mainStore, project_uuid);
     }
 
     render() {
@@ -45,12 +47,13 @@ class NavigationTabsWrapper extends React.Component {
         this.scansSubscriber.close();
         this.filesSubscriber.close();
         this.notificationsSubscriber.close();
-        this.scopesSubscriber.close();
+        this.this.scopesSubscriber.close();
+        this.this.credsSubscriber.close();
     }
 }
 
 NavigationTabsWrapper.contextTypes = {
-    store: React.PropTypes.object
+    store: PropTypes.object
 }
 
 export default NavigationTabsWrapper;
