@@ -1,6 +1,12 @@
 import _ from 'lodash'
 import React from 'react'
-import { Modal, Button, Popup } from 'semantic-ui-react'
+import {
+	Modal,
+	Button,
+	Popup,
+	Message,
+	Divider
+} from 'semantic-ui-react'
 
 
 import CustomOptions from './CustomOptions.jsx'
@@ -20,19 +26,25 @@ class InnerModal extends React.Component {
 		this.onInputChange = this.onInputChange.bind(this);
 	}
 
-	componentWillReceiveProps(nextProps) {
-		var inputs = {};
-		_.map(nextProps.task.available_options, (x) => {
-			inputs[x.name] = {
-				type: x.type,
-				value: x.default_value,
-				default_value: x.default_value,
-			};				
-		});
+	shouldComponentUpdate(nextProps, nextState) {
+		return ((!_.isEqual(nextProps, this.props)) || (!_.isEqual(nextState, this.state)));
+	}
 
-		this.setState({
-			inputs: inputs
-		});
+	componentDidUpdate(prevProps) {
+		if (!_.isEqual(prevProps, this.props)) {
+			let inputs = {};
+			_.map(this.props.task.available_options, (x) => {
+				inputs[x.name] = {
+					type: x.type,
+					value: x.default_value,
+					default_value: x.default_value,
+				};				
+			});
+
+			this.setState({
+				inputs: inputs
+			});
+		}		
 	}
 
 	startTask() {
@@ -72,7 +84,7 @@ class InnerModal extends React.Component {
 
 	onInputChange(name, value) {
 		// TODO: change this
-		var inputs = this.state.inputs;
+		let inputs = JSON.parse(JSON.stringify(this.state.inputs));
 		inputs[name].value = value;
 
 		this.setState({
@@ -81,13 +93,15 @@ class InnerModal extends React.Component {
 	}
 
 	render() {
-		const startButtons = _.map(this.props.task.preformed_options, (x) => {
-			var options = [];
+		const { task, open, openModal, closeModal } = this.props;
+
+		const startButtons = _.map(task.preformed_options, (x) => {
+			let options = [];
 			_.forOwn(x.options, (value, key) => {
 				options.push(<div key={key}><strong>{key}:</strong> {value}</div>);
 			});
 
-			var button = (
+			let button = (
 				<span>
 					<Button size="small"
 					        id={"button_" + x.name}
@@ -105,16 +119,41 @@ class InnerModal extends React.Component {
 		});
 
 		return (
-			<Modal open={this.props.open} onOpen={this.props.openModal} onClose={this.props.closeModal} >
+			<Modal open={open} onOpen={openModal} onClose={closeModal} >
 				<Modal.Header>
 					Choose one of the prepared options or create your own
 				</Modal.Header>
 
 				<Modal.Content>
 					{startButtons}
-
-					<br/>
-					<br/>
+					<Divider hidden />
+					{
+						task.help && 
+						task.help.map((help_notice) => {
+							if (help_notice.condition === true) {
+								if (help_notice.type == 'warning') {
+									return (
+										<Message
+											warning
+											key={task.help.indexOf(help_notice)}
+										>
+											{help_notice.text}
+										</Message>
+									);
+								}
+								else {
+									return (
+										<Message
+											key={task.help.indexOf(help_notice)}
+										>
+											{help_notice.text}
+										</Message>
+									);
+								}
+							}
+						})
+					}
+					<Divider hidden />
 					<CustomOptions inputs={this.state.inputs}
 								   startTaskHandler={this.startTask}
 								   onInputChange={this.onInputChange} />					
