@@ -7,7 +7,7 @@ import Tasks from '../../common/tasks/Tasks.jsx'
 import IPsSocketioEventsEmitter from '../../redux/ips/IPsSocketioEventsEmitter.js'
 
 
-class ProjectDetails extends React.Component {
+class IPsList extends React.Component {
 	constructor(props) {
 		super(props);
 
@@ -17,12 +17,12 @@ class ProjectDetails extends React.Component {
   		{
   			this.emitter.requestTasksByIps(this.props.ips.data.map((ip) => {
   				return ip.ip_address;
-  			}), this.props.project.project_uuid);
+  			}), this.props.project_uuid);
   		}		
 	}
 
   	shouldComponentUpdate(nextProps) {
-  		return (!_.isEqual(nextProps, this.props));
+  		return !_.isEqual(nextProps, this.props);
   	}
 
   	componentDidUpdate(prevProps) {
@@ -38,24 +38,24 @@ class ProjectDetails extends React.Component {
   		{
   			this.emitter.requestTasksByIps(ips.data.map((ip) => {
   				return ip.ip_address;
-  			}), project.project_uuid);
+  			}), project_uuid);
   		}
   	}
 
 	render() {
-		let { ips, hosts, project, filters, applyFilters, setLoaded, renewIps } = this.props;
+		let { ips, tasks, project_uuid, filters, applyFilters, setLoaded, renewIps } = this.props;
 
 		return (
 			<div>
-				<Tasks tasks={this.props.tasks} />
+				<Tasks tasks={tasks} />
 				<br/>
 
-				<TitleButtonsWithHandlers scopes={ips.data}
-									      project={project} 
-									      filters={this.props.filters} />
+				<TitleButtonsWithHandlers
+					project_uuid={project_uuid} 
+					filters={filters} />
 
 				<IPTableTracked ips={ips}
-								project_uuid={project.project_uuid}
+								project_uuid={project_uuid}
 								applyFilters={applyFilters}
 								filters={filters}
 								setLoaded={setLoaded}
@@ -65,4 +65,4 @@ class ProjectDetails extends React.Component {
 	}
 }
 
-export default ProjectDetails;
+export default IPsList;
