@@ -3,15 +3,14 @@ import React from 'react'
 import {
     Link
 } from 'react-router-dom'
-import classnames from 'classnames';
 import { Tab, Menu, Icon } from 'semantic-ui-react'
 
 import ScopeSetupWrapper from '../scope_setup/components/ScopeSetupWrapper.js'
-import ProjectDetailsWrapper from '../ips_list/components/ProjectDetailsWrapper.js'
+import IPsListWrapper from '../ips_list/components/IPsListWrapper.js'
 import HostsListWrapper from '../hosts_list/components/HostsListWrapper.js'
-import TasksTabWrapper from '../tasks_tab/components/TasksTabWrapper.js'
 
-import Perf from 'react-addons-perf'
+
+// import Perf from 'react-addons-perf'
 
 
 class NavigationTabs extends React.Component {
@@ -26,16 +25,16 @@ class NavigationTabs extends React.Component {
 
 	}
 
-	componentDidMount() {
-		setTimeout(() => {
-		  Perf.start();
-		  setTimeout(() => {
-			Perf.stop();
-			const measurements = Perf.getLastMeasurements();
-			Perf.printInclusive(measurements);
-		  }, 30000);
-		}, 5000);
-	  }
+	// componentDidMount() {
+	// 	setTimeout(() => {
+	// 	  Perf.start();
+	// 	  setTimeout(() => {
+	// 		Perf.stop();
+	// 		const measurements = Perf.getLastMeasurements();
+	// 		Perf.printInclusive(measurements);
+	// 	  }, 30000);
+	// 	}, 5000);
+	//   }
 
 	shouldComponentUpdate(nextProps, nextState) {
 		return ((!_.isEqual(nextProps, this.props)) || (!_.isEqual(nextState, this.state)));
@@ -56,7 +55,7 @@ class NavigationTabs extends React.Component {
 			},
 			{
 				menuItem: 'IP List',
-				render: (() => <ProjectDetailsWrapper project_uuid={this.project_uuid} />)
+				render: (() => <IPsListWrapper project_uuid={this.project_uuid} />)
 			},
 			{
 				menuItem: 'Hosts List',
