@@ -21,11 +21,45 @@ class FilesSocketioEventsEmitter {
         });
     }
 
-    renewStats(project_uuid, hostname) {
-        this.connector.emit('files:stats:get', {
+    renewCount(project_uuid) {
+        this.connector.emit('files:count:get', {
             "project_uuid": project_uuid
         });
-    }    
+    }
+
+    requestStatsHost(project_uuid, host_ids) {
+        this.connector.emit('files:stats:get:host', {
+            "project_uuid": project_uuid,
+            "host_ids": host_ids
+        });
+    }
+
+    requestStatsIPs(project_uuid, ip_ids) {
+        this.connector.emit('files:stats:get:ip', {
+            "project_uuid": project_uuid,
+            "ip_ids": ip_ids
+        });
+    }
+
+    requestFilesHosts(project_uuid, host, port_number, limit, offset) {
+        this.connector.emit('files:get:hosts', {
+            "project_uuid": project_uuid,
+            "host": host,
+            "port_number": port_number,
+            "limit": limit,
+            "offset": offset
+        });
+    }
+
+    requestFilesIps(project_uuid, ip, port_number, limit, offset) {
+        this.connector.emit('files:get:ips', {
+            "project_uuid": project_uuid,
+            "ip": ip,
+            "port_number": port_number,
+            "limit": limit,
+            "offset": offset
+        });
+    }
 }
 
 export default FilesSocketioEventsEmitter;
