@@ -1,9 +1,44 @@
-export const RENEW_FILES_STATS = 'RENEW_FILES_STATS'
+export const RENEW_TOTAL_AMOUNT = 'RENEW_TOTAL_AMOUNT'
+export const ADD_STATS_HOSTS = 'ADD_STATS_HOSTS'
+export const ADD_STATS_IPS = 'ADD_STATS_IPS'
+export const ADD_FILES_IPS = 'ADD_FILES_IPS'
+export const ADD_FILES_HOSTS = 'ADD_FILES_HOSTS'
 
 
-export function renewFilesStats(message, current_project_uuid) {
+export function renewTotalAmount(message, current_project_uuid) {
 	return {
-		type: RENEW_FILES_STATS,
+		type: RENEW_TOTAL_AMOUNT,
+		current_project_uuid: current_project_uuid,
+		message
+	}
+}
+
+export function addStatsIps(message, current_project_uuid) {
+	return {
+		type: ADD_STATS_IPS,
+		current_project_uuid: current_project_uuid,
+		message
+	}
+}
+
+export function addStatsHosts(message, current_project_uuid) {
+	return {
+		type: ADD_STATS_HOSTS,
+		current_project_uuid: current_project_uuid,
+		message
+	}
+}
+export function addFilesHosts(message, current_project_uuid) {
+	return {
+		type: ADD_FILES_HOSTS,
+		current_project_uuid: current_project_uuid,
+		message
+	}
+}
+
+export function addFilesIps(message, current_project_uuid) {
+	return {
+		type: ADD_FILES_IPS,
 		current_project_uuid: current_project_uuid,
 		message
 	}
