@@ -9,6 +9,7 @@ import file_reduce from './files/reducers';
 import filter_reduce from './filters/reducers';
 import notification_reduce from './notifications/reducers';
 import scope_reduce from './scopes/reducers';
+import creds_reduce from './creds/reducers';
 
 
 const all_reducers = combineReducers({
@@ -20,7 +21,8 @@ const all_reducers = combineReducers({
 	files: file_reduce,
 	filters: filter_reduce,
 	notifs: notification_reduce,
-	scopes: scope_reduce
+	scopes: scope_reduce,
+	creds: creds_reduce
 });
 
 export default all_reducers;
