@@ -1,7 +1,7 @@
 export const RENEW_SCANS_STATS = 'RENEW_SCANS_STATS'
 
 
-export function renewStats(message, current_project_uuid) {
+export function renewCount(message, current_project_uuid) {
     return { 
         type: RENEW_SCANS_STATS,
         current_project_uuid: current_project_uuid,
