@@ -1,18 +1,14 @@
-""" Simplt keeps handlers for scans (open ports) """
-
-
 class ScanHandlers(object):
-    """ This class is really simple to read """
-
     def __init__(self, socketio, scan_manager):
         self.socketio = socketio
         self.scan_manager = scan_manager
 
+        self.register_handlers()
+
     def register_handlers(self):
-        """ Register the single handler for returning all scans """
         @self.socketio.on('scans:stats:get', namespace='/scans')
         async def _cb_handle_custom_event(sio, msg):
-            """ When received this message, send back all the scans """
+            """ When received this message, send stats of the scans """
             project_uuid = int(msg.get('project_uuid', None))
 
             await self.send_stats_back(project_uuid)
@@ -28,26 +24,3 @@ class ScanHandlers(object):
             },
             namespace='/scans'
         )
-
-    async def notify_on_updated_scans(self, new_ips, project_uuid=None):
-        """ Finds all scans for the project and sends them back """
-        print("Sending back new_ips {} , {}".format(new_ips, project_uuid))
-
-        if new_ips:
-            await self.socketio.emit(
-                'ips:updated', {
-                    'status': 'success',
-                    'project_uuid': project_uuid,
-                    'updated_ips': new_ips
-                },
-                namespace='/ips'
-            )
-
-            await self.socketio.emit(
-                'hosts:updated:ips', {
-                    'status': 'success',
-                    'project_uuid': project_uuid,
-                    'updated_ips': new_ips
-                },
-                namespace='/hosts'
-            )
