@@ -1,4 +1,3 @@
-""" Module keeps class of scope handlers """
 import asyncio
 from netaddr import IPNetwork
 
@@ -11,10 +10,12 @@ class IPHandlers(object):
         self.socketio = socketio
         self.scope_manager = scope_manager
 
+        self.register_handlers()
+
     def register_handlers(self):
         @self.socketio.on('ips:part:get', namespace='/ips')
         async def _cb_handle_scopes_get(sio, msg):
-            """ When received this message, send back all the scopes """
+            """ When received this message, send back all the ips """
             project_uuid = int(msg.get('project_uuid', None))
             ip_page = msg.get('ip_page', 0)
             ip_page_size = msg.get('ip_page_size', 12)
@@ -26,11 +27,9 @@ class IPHandlers(object):
 
         @self.socketio.on('ips:single:get', namespace='/ips')
         async def _cb_handle_scope_single_get(sio, msg):
-            """ When received this message, send back all the scopes """
+            """ When received this message, send back a single ip address """
             project_uuid = int(msg.get('project_uuid', None))
             ip_address = msg.get('ip_address')
-
-            print("Getting single ip", ip_address)
 
             await self.send_ips_back({'ip': [ip_address]}, sio, project_uuid)
 
@@ -92,6 +91,7 @@ class IPHandlers(object):
             """ When received this message, send back all the tasks """
             project_uuid = int(msg.get('project_uuid', None))
             ips = msg.get('ips', None)
+
             await self.send_tasks_back_filtered(project_uuid, ips=ips)
 
     async def send_tasks_back_filtered(self, project_uuid, ips=None, hosts=None):
@@ -126,41 +126,24 @@ class IPHandlers(object):
     ):
         """ Collects all relative hosts and ips from
         the manager and sends them back """
-
-        ips = self.scope_manager.get_ips(
+        ips = self.scope_manager.get_ips_with_ports(
             filters, project_uuid, ip_page, ip_page_size)
 
-        if sio is None:
-            await self.socketio.emit(
-                'ips:part:set', {
-                    'status': 'success',
-                    'project_uuid': project_uuid,
-                    'ips': {
-                        'page': ip_page,
-                        'page_size': ip_page_size,
-                        'data': ips['ips'],
-                        'selected_ips': ips['selected_ips'],
-                        'total_db_ips': ips['total_db_ips']
-                    }
-                },
-                namespace='/ips'
-            )
-        else:
-            await self.socketio.emit(
-                'ips:part:set', {
-                    'status': 'success',
-                    'project_uuid': project_uuid,
-                    'ips': {
-                        'page': ip_page,
-                        'page_size': ip_page_size,
-                        'data': ips['ips'],
-                        'selected_ips': ips['selected_ips'],
-                        'total_db_ips': ips['total_db_ips']
-                    }
-                },
-                room=sio,
-                namespace='/ips'
-            )
+        await self.socketio.emit(
+            'ips:part:set', {
+                'status': 'success',
+                'project_uuid': project_uuid,
+                'ips': {
+                    'page': ip_page,
+                    'page_size': ip_page_size,
+                    'data': ips['ips'],
+                    'selected_ips': ips['selected_ips'],
+                    'total_db_ips': ips['total_db_ips']
+                }
+            },
+            room=sio,
+            namespace='/ips'
+        )
 
 
 class HostHandlers(object):
@@ -168,6 +151,8 @@ class HostHandlers(object):
     def __init__(self, socketio, scope_manager):
         self.socketio = socketio
         self.scope_manager = scope_manager
+
+        self.register_handlers()
 
     def register_handlers(self):
         @self.socketio.on('hosts:part:get', namespace='/hosts')
@@ -222,12 +207,12 @@ class HostHandlers(object):
             comment = msg['comment']
             project_uuid = msg['project_uuid']
 
-            result = self.scope_manager.update_scope(
+            update_result = self.scope_manager.update_scope(
                 scope_id=host_id, comment=comment, scope_type='host'
             )
-            target = result['target']
+            target = update_result['target']
 
-            if result["status"] == "success":
+            if update_result["status"] == "success":
                 await self.socketio.emit(
                     'hosts:update:back',
                     {
@@ -248,10 +233,10 @@ class HostHandlers(object):
                     project_uuid=project_uuid
                 )
             else:
-                result['project_uuid'] = project_uuid
+                update_result['project_uuid'] = project_uuid
                 await self.socketio.emit(
                     'hosts:update:back',
-                    result,
+                    update_result,
                     namespace='/hosts'
                 )
 
@@ -270,6 +255,7 @@ class HostHandlers(object):
             """ When received this message, send back all the tasks """
             project_uuid = int(msg.get('project_uuid', None))
             hosts = msg.get('hosts', None)
+
             await self.send_tasks_back_filtered(project_uuid, hosts=hosts)
 
 
@@ -312,55 +298,36 @@ class HostHandlers(object):
         """ Collects all relative hosts and ips from
         the manager and sends them back """
 
-        hosts = self.scope_manager.get_hosts(
+        hosts = self.scope_manager.get_hosts_with_ports(
             filters, project_uuid, host_page, host_page_size)
-        if sio is None:
-            await self.socketio.emit(
-                'hosts:part:set', {
-                    'status': 'success',
-                    'project_uuid': project_uuid,
-                    'hosts': {
-                        'page': host_page,
-                        'page_size': host_page_size,
-                        'data': hosts['hosts'],
-                        'total_db_hosts': hosts['total_db_hosts'],
-                        'selected_hosts': hosts['selected_hosts']
-                    }
-                },
-                namespace='/hosts'
-            )
-        else:
-            await self.socketio.emit(
-                'hosts:part:set', {
-                    'status': 'success',
-                    'project_uuid': project_uuid,
-                    'hosts': {
-                        'page': host_page,
-                        'page_size': host_page_size,
-                        'data': hosts['hosts'],
-                        'total_db_hosts': hosts['total_db_hosts'],
-                        'selected_hosts': hosts['selected_hosts']
-                    }
-                },
-                room=sio,
-                namespace='/hosts'
-            )
+
+        await self.socketio.emit(
+            'hosts:part:set', {
+                'status': 'success',
+                'project_uuid': project_uuid,
+                'hosts': {
+                    'page': host_page,
+                    'page_size': host_page_size,
+                    'data': hosts['hosts'],
+                    'total_db_hosts': hosts['total_db_hosts'],
+                    'selected_hosts': hosts['selected_hosts']
+                }
+            },
+            room=sio,
+            namespace='/hosts'
+        )
 
 
 class ScopeHandlers(object):
-    """ Registers all handlers related to scopes """
-
     def __init__(self, socketio, scope_manager):
         self.socketio = socketio
         self.scope_manager = scope_manager
 
-    def register_handlers(self):
-        """ Register all handlers """
+        self.register_handlers()
 
-        ip = IPHandlers(self.socketio, self.scope_manager)
-        ip.register_handlers()
-        host = HostHandlers(self.socketio, self.scope_manager)
-        host.register_handlers()
+    def register_handlers(self):
+        IPHandlers(self.socketio, self.scope_manager)
+        HostHandlers(self.socketio, self.scope_manager)
 
         @self.socketio.on('scopes:create', namespace='/scopes')
         async def _cb_handle_scope_create(sio, msg):
@@ -375,7 +342,6 @@ class ScopeHandlers(object):
             error_text = ""
 
             for scope in scopes:
-                added = False
                 # Create newly added scope
                 if scope['type'] == 'hostname':
                     create_result = self.scope_manager.create_host(
